Rename private utils file
