+++ lib/src/react_client/dart_interop_statics.dart
@@ -13,7 +13,7 @@ import 'package:react/react_client/react_interop.dart';
 import 'package:react/react_client/zone.dart';
 
 import 'package:react/src/context.dart';
-import 'package:react/src/react_client/utils.dart';
+import 'package:react/src/react_client/private_utils.dart';
 import 'package:react/src/typedefs.dart';
 
 /// The static methods that proxy JS component lifecycle methods to Dart components.
