+++ lib/src/react_client/private_utils.dart
